Move Homepage page to correct folder
